+++ app/ChatRoom.tsx
@@ -10,7 +10,8 @@ import { IconSymbol } from '@/components/ui/IconSymbol';
 import { useThemeColor } from '@/hooks/useThemeColor';
 import * as ImagePicker from 'expo-image-picker';
 import * as ImageManipulator from 'expo-image-manipulator';
-import { MediaAttachment, Chat, Message, User } from '@/types/types';
+import { MediaAttachment, Message, User } from '@/types/types';
+import styles from '@/styles/ChatRoomStyles';
 
 export default function ChatRoomScreen() {
   const { chatId } = useLocalSearchParams<{ chatId: string }>();
@@ -218,72 +219,3 @@ export default function ChatRoomScreen() {
     </KeyboardAvoidingView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  centerContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-  },
-  messagesContainer: {
-    padding: 10,
-    flexGrow: 1,
-  },
-  emptyContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  inputContainer: {
-    flexDirection: 'row',
-    padding: 10,
-    alignItems: 'flex-end',
-    borderTopWidth: 1,
-  },
-  input: {
-    flex: 1,
-    borderWidth: 1,
-    borderRadius: 20,
-    padding: 10,
-    maxHeight: 100,
-  },
-  sendButton: {
-    marginLeft: 10,
-    marginBottom: 5,
-    width: 40,
-    height: 40,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  disabledButton: {
-    opacity: 0.5,
-  },
-  mediaButton: {
-    marginRight: 10,
-    marginBottom: 5,
-    width: 40,
-    height: 40,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  mediaPreviewContainer: {
-    maxHeight: 100,
-    marginBottom: 10,
-    paddingHorizontal: 10,
-  },
-  mediaThumbnail: {
-    width: 80,
-    height: 80,
-    borderRadius: 8,
-    marginRight: 8,
-  },
-});
