+++ components/ThemedText.tsx
@@ -1,6 +1,6 @@
-import { Text, type TextProps, StyleSheet } from 'react-native';
-
+import { Text, type TextProps} from 'react-native';
 import { useThemeColor } from '@/hooks/useThemeColor';
+import styles from '@/styles/ThemedTextStyles';
 
 export type ThemedTextProps = TextProps & {
   lightColor?: string;
@@ -33,29 +33,4 @@ export function ThemedText({
       {...rest}
     />
   );
-}
-
-const styles = StyleSheet.create({
-  default: {
-    fontSize: 16,
-    lineHeight: 24,
-  },
-  defaultSemiBold: {
-    fontSize: 16,
-    lineHeight: 24,
-    fontWeight: '600',
-  },
-  title: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    lineHeight: 32,
-  },
-  subtitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  link: {
-    lineHeight: 30,
-    fontSize: 16
-  },
-});
+}
