+++ components/UserListItem.tsx
@@ -3,6 +3,7 @@ import { View, StyleSheet, Pressable } from 'react-native';
 import { ThemedText } from './ThemedText';
 import { Avatar } from './Avatar';
 import { User } from '@/hooks/useUser';
+import styles from '../styles/UserListItemStyles';
 
 interface UserListItemProps {
   user: User;
@@ -32,25 +33,3 @@ export function UserListItem({ user, onSelect, isSelected }: UserListItemProps) 
     </Pressable>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 12,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    borderBottomColor: '#E1E1E1',
-  },
-  selectedContainer: {
-    backgroundColor: 'rgba(0, 122, 255, 0.1)',
-  },
-  infoContainer: {
-    marginLeft: 12,
-    flex: 1,
-  },
-  statusText: {
-    fontSize: 14,
-    color: '#8F8F8F',
-    marginTop: 4,
-  },
-});
