+++ styles/ThemedTextStyles.ts
@@ -1,0 +1,28 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    default: {
+      fontSize: 16,
+      lineHeight: 24,
+    },
+    defaultSemiBold: {
+      fontSize: 16,
+      lineHeight: 24,
+      fontWeight: '600',
+    },
+    title: {
+      fontSize: 32,
+      fontWeight: 'bold',
+      lineHeight: 32,
+    },
+    subtitle: {
+      fontSize: 20,
+      fontWeight: 'bold',
+    },
+    link: {
+      lineHeight: 30,
+      fontSize: 16
+    },
+  });
+
+export default styles;
